// src/app/shared/model/update-master-cust/update-cust-personal-detail-obj.model.ts
export class UpdateCustPersonalDetailObj {
    CustId: any;
    CustPersonalId: any;
    CustFullName: any;
    MrMaritalStatCode: any;
    MrNationalityCode: any;
    Country: any;
    CountryName: any;
    MrEducationCode: any;
    MrReligionCode: any;
    FamilyCardNo: any;
    NoOfResidence: any;
    NoOfDependents: any;
    CustomerGroupParentCustId: any;
    CustomerGroupParentCustName: any;
    IsAffiliationWithMf: any;
    IsVip: any;
    VipNotes: any;
    MobilePhnNo1: any;
    MobilePhnNo2: any;
    Email1: any;
    Email2: any;
    RowVersionCust: any;
    RowVersionCustPersonal: any;
    RowVersionCustGrp: any;
}